+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.1.2] (2020-05-16)
+
 - Add `Get-TimeSinceReboot` function
 - Add `Read-Path` function
 - Add `Set-EnvironmentVariable` function
@@ -22,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Add `New-Directory`, `Measure-Directory` and `Remove-Directory` functions
 - Add `New-Credential` function
 
-[Unreleased]: https://github.com/kitforbes/UtilitiesPS/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/kitforbes/UtilitiesPS/compare/v0.1.2...HEAD
+[v0.1.2]: https://github.com/kitforbes/UtilitiesPS/compare/v0.1.1...v0.1.2
 [v0.1.1]: https://github.com/kitforbes/UtilitiesPS/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/kitforbes/UtilitiesPS/compare/1e5f30c...v0.1.0